Add files via upload
A Terraform module to create AWS resources which are used to automatically take backup of all the parameters residing on AWS SSM Parameter Store in JSON format and store it on AWS S3 bucket using AWS Lambda function based on Python. It is executed daily via AWS CloudWatch or AWS EventBridge.